set "show_header" to unchecked and moved cloneNodeFlow
